add scaling picture to django
